requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
httpie==2.0.0
httplib2==0.9.2
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
isc==2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pbr==5.4.4
pew==1.2.0
pexpect==4.2.1
Pillow==7.1.1
pipenv==2018.11.26
ply==3.11
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.5.2
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-pptx==0.6.18
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
sqlparse==0.4.1
stevedore==1.31.0
stripe==2.44.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.3
virtualenv==20.0.1
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wadllib==1.3.2
xkit==0.0.0
XlsxWriter==1.2.8
zipp==2.2.0
zope.interface==4.3.2
